Allow binding input processors and interactions
Interactions do not currently work properly because we are only looking at whether the "start" phase was triggered, not "performed". E.g. a multiTap() will trigger "start" when the button is pressed, and only "performed" when/if the button is pressed the required number of times within the designated time period.