Load VR Entities from YAML; refactor Devices spec
Devices now have a separate vrEntity node in the YAML which contains bindings and offsets for position and rotation, scale, and model . Removed `mountTo` and the two hand mount gameObjects in the Scene.

The `model` field allows users to choose one of Unity's six primitives to use to represent the Device.

This means that the user can define IMUs, set rotation offsets for them, and bind objects (cuboids or otherwise) to them in VR space.