Explicitly send static IMU values while gathering samples
Previously, IMU devices would not send any update to device state (i.e. to pos/rot/accel/gyro) until 50 samples of pos/rot had been taken, to allow smoothing.

It's possible that this is the source of the problem here: https://github.com/MichaelJW/DorsalVR/issues/7 - by sending nothing for so long, during Dolphin's calibration period, it may be the case that it's accidentally screwing up the calibration.

I can't reproduce the problem on my end, so this is a bit of a stab in the dark. But I have at least verified it doesn't break the build for me.